requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.6.0
urllib3==1.26.8
